[UrlReference] Fix conversion errors for collections in GraphNodes.
